refactor(frontend): split model page into nested layout routes
Each tab is now a proper URL (/models/[slug]/edit, /models/[slug]/activity)
instead of in-page state. The layout load runs once per slug; switching
tabs does not re-fetch data.

- +layout.ts: shared model load (renamed from +page.ts)
- +layout.svelte: header, tab bar as <a> links, footer
- +page.svelte: detail tab content only
- edit/+page.svelte: edit form; resets via invalidateAll() after save
- activity/+page.svelte: activity sections
